+++ src/browser/types.ts
@@ -7,6 +7,7 @@ import { MoveCursorToCommand } from "./commands/moveCursorTo";
 import { OpenAndWaitCommand } from "./commands/openAndWait";
 import type { Callstack } from "./history/callstack";
 import { Test, Hook } from "../test-reader/test-object";
+import type { CaptureSnapshotOptions, CaptureSnapshotResult } from "./commands/captureDomSnapshot";
 import { TestplaneQueries, TestplaneQueriesChainable, TestplaneQueriesSync } from "./queries";
 
 export const BrowserName = {
@@ -154,6 +155,16 @@ declare global {
             switchToRepl: (this: WebdriverIO.Browser, ctx?: Record<string, unknown>) => Promise<void>;
 
             clearSession: (this: WebdriverIO.Browser) => Promise<void>;
+
+            unstable_captureDomSnapshot(
+                this: WebdriverIO.Browser,
+                options?: Partial<CaptureSnapshotOptions>,
+            ): Promise<CaptureSnapshotResult>;
+            unstable_captureDomSnapshot(
+                this: WebdriverIO.Browser,
+                selector: string,
+                options?: Partial<CaptureSnapshotOptions>,
+            ): Promise<CaptureSnapshotResult>;
         }
 
         interface Element {
@@ -190,6 +201,11 @@ declare global {
             assertView: AssertViewElementCommand;
 
             moveCursorTo: MoveCursorToCommand;
+
+            unstable_captureDomSnapshot(
+                this: WebdriverIO.Element,
+                options?: Partial<CaptureSnapshotOptions>,
+            ): Promise<CaptureSnapshotResult>;
         }
     }
 }