+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [8.29.5](https://github.com/gemini-testing/testplane/compare/v8.29.4...v8.29.5) (2025-06-24)
+
+
+### Bug Fixes
+
+* use webdriverio connection retries ([194d2f7](https://github.com/gemini-testing/testplane/commit/194d2f77a5bb2c7f2494b6d7605efc581971eb91))
+
+### [8.29.4](https://github.com/gemini-testing/testplane/compare/v8.29.3...v8.29.4) (2025-06-23)
+
+
+### Bug Fixes
+
+* fix missing steps when time travel is on ([#1102](https://github.com/gemini-testing/testplane/issues/1102)) ([ec5e174](https://github.com/gemini-testing/testplane/commit/ec5e17463a1b660e46beb16c9b4ac200a8f4e872))
+* use RAF timing in rrweb snapshots ([#1103](https://github.com/gemini-testing/testplane/issues/1103)) ([804a131](https://github.com/gemini-testing/testplane/commit/804a1318358508a6b10468375fc6ea11e5c2d2fe))
+
+### [8.29.3](https://github.com/gemini-testing/testplane/compare/v8.29.2...v8.29.3) (2025-06-18)
+
+
+### Features
+
+* add ability to capture DOM snapshots of specific elements ([#1098](https://github.com/gemini-testing/testplane/issues/1098)) ([8836a50](https://github.com/gemini-testing/testplane/commit/8836a503c8ae9b49e20ba5d0b18c83e806c5eb68))
+* implement captureDomSnapshot command ([#1097](https://github.com/gemini-testing/testplane/issues/1097)) ([3a1b1da](https://github.com/gemini-testing/testplane/commit/3a1b1da10f844597912b2b0db96c881bd5a6879c))
+
 ### [8.29.2](https://github.com/gemini-testing/testplane/compare/v8.29.1...v8.29.2) (2025-05-20)